refactor: extract logging into log_manager.{hpp,cpp}
Moves InitLogger/Log/LogSEH and the g_log_file/g_logPath globals
into their own translation unit, plus a new CloseLogger() that
consolidates the duplicated cleanup from both wWinMain branches.

- cli_args_debugger.cpp: 2101 -> 1964 lines (-137)
- extern g_log_file / g_logPath kept for tests/logging_tests.cpp
- CMakeLists, tests/CMakeLists, build.yml, validate.sh, README updated
