+++ .github/workflows/build.yml
@@ -30,7 +30,7 @@ jobs:
       shell: cmd
       run: |
         cl /EHsc /std:c++20 /permissive- /I. /DUNICODE /D_UNICODE /GS /sdl ^
-           cli_args_debugger.cpp seh_wrapper.cpp qrcodegen.cpp ^
+           cli_args_debugger.cpp log_manager.cpp seh_wrapper.cpp qrcodegen.cpp ^
            /Fe:build\cloud-streaming-args-debugger.exe ^
            /Fo:obj\ ^
            /link d3d11.lib d3dcompiler.lib d2d1.lib dwrite.lib ole32.lib avrt.lib user32.lib shell32.lib gdi32.lib propsys.lib winmm.lib psapi.lib
